Shareef/another button refactor (#25)
