+++ lib/views/mobile/settings.dart
@@ -74,7 +74,7 @@ class _SettingsState extends State<Settings> {
                           ],
                         ),
                         ButtonWithBackground(
-                            handleSelectFolder: handleSelectFile,
+                            handleSelectFolder: (){},
                             key:Key(SETTINGS_SCREEN_SELECT_A_FOLDER_BUTTON)
                         ),
                       Button(
